chore: move cookie package to utils package
